Update subject naming, heuristic
Current workflow grabs the file path & name by the site. This causes a
conflict with the LHS dataset as it is split between A and B. For now,
the fix is to rename the subjects file appropriately and calling the
correct heuristic.

NA: Look into stripping specific phase for LHS datasets.